refactor(lore): Create cmd/lore/lore Application, delete internal/starlark and internal/lore
Phase 2 of the consolidate-starlark-runtime plan.

Move lore-specific code into cmd/lore/lore/ following the cmd/star/star/
Application pattern. Eliminate the internal/starlark.Runtime wrapper —
all consumers now use op.StarlarkRuntime directly.

Moves:
- internal/starlark/ (lore-specific) -> cmd/lore/lore/
- internal/lore/ -> cmd/lore/lore/

Deletes:
- internal/starlark/runtime.go (wrapper, superseded by op.StarlarkRuntime)
- internal/starlark/plan.go (empty)
- internal/starlark/receiver.go (empty)

Import updates:
- cmd/lore/main.go, cmd/docgen/main.go, internal/writ/commands.go,
  internal/writ/graph_builder.go, internal/e2e/onboard_test.go
- internal/starlarktest/runner.go now uses op.NewStarlarkRuntime directly
