+++ cmd/docgen/main.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/NobleFactor/devlore-cli/internal/lore"
+	"github.com/NobleFactor/devlore-cli/cmd/lore/lore"
 	"github.com/NobleFactor/devlore-cli/internal/tools/docgen"
 	"github.com/NobleFactor/devlore-cli/internal/writ"
 )
