+++ cmd/lore/lore/builder_test.go
