+++ cmd/lore/lore/commands_test.go
