+++ cmd/lore/lore/onboard/onboard.go
