+++ cmd/lore/lore/package_context.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: SSPL-1.0
 // Copyright (c) 2025-2026 Noble Factor. All rights reserved.
 
-package starlark
+package lore
 
 import (
 	"fmt"
