+++ cmd/lore/lore/testdata/load_test.star
