--- internal/starlark/receiver.go
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: SSPL-1.0
-// Copyright (c) 2025-2026 Noble Factor. All rights reserved.
-
-package starlark
